+++ app/Http/Controllers/Admin/NoseMasterController.php
@@ -148,7 +148,7 @@ class NoseMasterController extends Controller
         $nose_obj = new nose_master();
         $nose_result = nose_master::find($nose_id);
 
-        if (isset($nose_result->nose_img) && !empty($nose_result->nose_img)) {
+        if (isset($nose_result->nose_img) && !empty($nose_result->nose_img) && file_exists(\public_path('uploads/noses/thumbnail/thumb_' . $nose_result->nose_img))) {
             $mime_type = $this->_base64_mime_type($nose_result->nose_img);
             $nose_result->nose_img = $mime_type . base64_encode(file_get_contents(\public_path('uploads/noses/thumbnail/thumb_' . $nose_result->nose_img)));
         }
@@ -163,14 +163,14 @@ class NoseMasterController extends Controller
      */
     private function _base64_mime_type($filename = "", $extention_only = false)
     {
+        $mime_types = config('mime_types');
         if (!$extention_only) {
-            $mime_types = config('mime_types');
-            if (in_array(strstr($filename, "."), array_flip($mime_types['mime_types']))) {
-                return "data:" . $mime_types['mime_types'][strstr($filename, ".")] . ";base64,";
+            if (in_array(strtolower(strstr($filename, ".")), array_flip($mime_types['mime_types']))) {
+                return "data:" . $mime_types['mime_types'][strtolower(strstr($filename, "."))] . ";base64,";
             }
         } else {
-            if (in_array($filename, array_flip($mime_types['mime_types']))) {
-                return "data:" . $mime_types['mime_types'][$filename] . ";base64,";
+            if (in_array(strtolower($filename), array_flip($mime_types['mime_types']))) {
+                return "data:" . $mime_types['mime_types'][strtolower($filename)] . ";base64,";
             }
         }
     }
@@ -200,7 +200,7 @@ class NoseMasterController extends Controller
             // Remove Old Uploaded Files From Folder
             if ($filehandle['status']) {
                 $nose_data = nose_master::find($update_id);
-                if (isset($nose_data) && !empty($nose_data) && !empty($nose_data->nose_img)) {
+                if (isset($nose_data) && !empty($nose_data) && !empty($nose_data->nose_img) && file_exists(\public_path("uploads/noses/$nose_data->nose_img")) && file_exists(\public_path("uploads/noses/thumbnail/thumb_$nose_data->nose_img"))) {
                     unlink(\public_path("uploads/noses/$nose_data->nose_img"));
                     unlink(\public_path("uploads/noses/thumbnail/thumb_$nose_data->nose_img"));
                 }
